Put server code into src directory, and extract socket code from index.js into seperate module
